+++ nodes/SeaAdmin/v1/actions/bases/delete_base/index.ts
@@ -1,0 +1,4 @@
+import { delete_base as execute } from './execute';
+import { description as description } from './description';
+
+export { description, execute };