+++ nodes/SeaAdmin/v1/actions/users/reset_user_password/index.ts
@@ -1,0 +1,4 @@
+import { reset_user_password as execute } from './execute';
+import { description as description } from './description';
+
+export { description, execute };